pyproject.toml
[project]
name = "ichatbio-data-handler-agent"
version = "1"
requires-python = ">=3.12"
dependencies = [
    "ichatbio-sdk~=0.2.1",
    "openai~=1.79.0",
    "pydantic~=2.11.4",
    "instructor~=1.8.2",
    "dotenv~=0.9.9",
    "pytest~=8.3.5",
    "pytest-asyncio~=1.0.0a1",
    "starlette~=0.47.3",
    "uvicorn~=0.34.3",
    "genson~=1.3.0",
    "jq~=1.10.0",
    "pytest_httpx~=0.35.0",
    "langchain~=1.0.0a4",
    "langchain_openai~=1.0.0a2"
]

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
log_cli = true
log_cli_level = "INFO"